requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pytesseract==0.3.13
python-dotenv==1.2.1
python-multipart==0.0.21
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
WTForms==3.2.1
